Move setters to backend
