+++ src/qibo/config.py
@@ -44,6 +44,34 @@ def raise_error(exception, message=None, args=None):
         raise exception(message)
 
 
+# Set the number of threads from the environment variable
+OMP_NUM_THREADS = None
+if "OMP_NUM_THREADS" not in os.environ:
+    import psutil
+    # using physical cores by default
+    cores = psutil.cpu_count(logical=False)
+    OMP_NUM_THREADS = cores
+else: # pragma: no cover
+    OMP_NUM_THREADS = int(os.environ.get("OMP_NUM_THREADS"))
+
+def get_threads():
+    """Returns number of threads."""
+    return OMP_NUM_THREADS
+
+def set_threads(num_threads):
+    """Set number of OpenMP threads.
+
+    Args:
+        num_threads (int): number of threads.
+    """
+    if not isinstance(num_threads, int): # pragma: no cover
+        raise_error(RuntimeError, "Number of threads must be integer.")
+    if num_threads < 1: # pragma: no cover
+        raise_error(RuntimeError, "Number of threads must be positive.")
+    global OMP_NUM_THREADS
+    OMP_NUM_THREADS = num_threads
+
+
 # Backend access
 from qibo import backend as K
 def set_computation_backend(backend="tensorflow"):
@@ -55,70 +83,57 @@ def set_computation_backend(backend="tensorflow"):
 
 set_computation_backend()
 
+os.environ["TF_CPP_MIN_LOG_LEVEL"] = str(LOG_LEVEL)
+import numpy as np
+import tensorflow as tf
+
+ARRAY_TYPES = (np.ndarray, tf.Tensor)
+NUMERIC_TYPES = (np.int, np.float, np.complex,
+                 np.int32, np.int64, np.float32,
+                 np.float64, np.complex64, np.complex128)
+
+
+# Define numpy and tensorflow matrices
+# numpy matrices are exposed to user via ``from qibo import matrices``
+# tensorflow matrices are used by native gates (``/tensorflow/gates.py``)
+from qibo.backend import matrices as _matrices
+matrices = _matrices.NumpyMatrices(np.complex128)
+tfmatrices = _matrices.TensorflowMatrices(tf.complex128)
+
+
+def set_precision(dtype='double'):
+    """Set precision for states and gates simulation.
+
+    Args:
+        dtype (str): possible options are 'single' for single precision
+            (complex64) and 'double' for double precision (complex128).
+    """
+    K.set_precision(dtype)
+    matrices.allocate_matrices()
+    tfmatrices.allocate_matrices()
+
+def get_precision():
+    """Get precision for states and gates simulation.
+
+    Returns:
+        A string with the precision name ('single', 'double').
+    """
+    return K.dtypes('STRING')
+
+
+def set_device(name):
+    K.set_device(name)
+    with tf.device(name):
+        tfmatrices.allocate_matrices()
+
+def get_device():
+    return K.active_device
+
 
 # Load backend specifics
 if BACKEND_NAME == "tensorflow":
-    os.environ["TF_CPP_MIN_LOG_LEVEL"] = str(LOG_LEVEL)
-    import numpy as np
-    import tensorflow as tf
-
-    ARRAY_TYPES = (np.ndarray, tf.Tensor)
-    NUMERIC_TYPES = (np.int, np.float, np.complex,
-                     np.int32, np.int64, np.float32,
-                     np.float64, np.complex64, np.complex128)
-
-    # Define numpy and tensorflow matrices
-    # numpy matrices are exposed to user via ``from qibo import matrices``
-    # tensorflow matrices are used by native gates (``/tensorflow/gates.py``)
-    from qibo.backend import matrices as _matrices
-    matrices = _matrices.NumpyMatrices(np.complex128)
-    tfmatrices = _matrices.TensorflowMatrices(tf.complex128)
-
-    # Set the number of threads from the environment variable
-    OMP_NUM_THREADS = None
-    if "OMP_NUM_THREADS" not in os.environ:
-        import psutil
-        # using physical cores by default
-        cores = psutil.cpu_count(logical=False)
-        OMP_NUM_THREADS = cores
-    else: # pragma: no cover
-        OMP_NUM_THREADS = int(os.environ.get("OMP_NUM_THREADS"))
-
-    def get_threads():
-        """Returns number of threads."""
-        return OMP_NUM_THREADS
-
-    def set_threads(num_threads):
-        """Set number of OpenMP threads.
-
-        Args:
-            num_threads (int): number of threads.
-        """
-        if not isinstance(num_threads, int): # pragma: no cover
-            raise_error(RuntimeError, "Number of threads must be integer.")
-        if num_threads < 1: # pragma: no cover
-            raise_error(RuntimeError, "Number of threads must be positive.")
-        global OMP_NUM_THREADS
-        OMP_NUM_THREADS = num_threads
-
     # Gate backends
     BACKEND = {'GATES': 'custom', 'EINSUM': None, 'STRING': 'custom'}
-
-    # Set devices recognized by tensorflow
-    DEVICES = {
-        'CPU': tf.config.list_logical_devices("CPU"),
-        'GPU': tf.config.list_logical_devices("GPU")
-    }
-    # set default device to GPU if it exists
-    if DEVICES['GPU']: # pragma: no cover
-        # case not tested by GitHub workflows because it requires a GPU
-        DEVICES['DEFAULT'] = DEVICES['GPU'][0].name
-    elif DEVICES['CPU']:
-        DEVICES['DEFAULT'] = DEVICES['CPU'][0].name
-    else: # pragma: no cover
-        # case not tested by GitHub workflows because it requires no device
-        raise_error(RuntimeError, "Unable to find Tensorflow devices.")
-
 
     def set_backend(backend='custom'):
         """Sets backend used to implement gates.
@@ -154,73 +169,6 @@ if BACKEND_NAME == "tensorflow":
             A string with the backend name.
         """
         return BACKEND['STRING']
-
-    def set_precision(dtype='double'):
-        """Set precision for states and gates simulation.
-
-        Args:
-            dtype (str): possible options are 'single' for single precision
-                (complex64) and 'double' for double precision (complex128).
-        """
-        if not ALLOW_SWITCHERS and dtype != DTYPES['STRING']:
-            warnings.warn("Precision should not be changed after allocating gates.",
-                          category=RuntimeWarning)
-        if dtype == 'single':
-            DTYPES['DTYPE'] = tf.float32
-            DTYPES['DTYPECPX'] = tf.complex64
-            DTYPES['NPTYPECPX'] = np.complex64
-        elif dtype == 'double':
-            DTYPES['DTYPE'] = tf.float64
-            DTYPES['DTYPECPX'] = tf.complex128
-            DTYPES['NPTYPECPX'] = np.complex128
-        else:
-            raise_error(RuntimeError, f'dtype {dtype} not supported.')
-        DTYPES['STRING'] = dtype
-        matrices.allocate_matrices()
-        tfmatrices.allocate_matrices()
-
-    def get_precision():
-        """Get precision for states and gates simulation.
-
-        Returns:
-            A string with the precision name ('single', 'double').
-        """
-        return DTYPES['STRING']
-
-    def set_device(device_name: str):
-        """Set default execution device.
-
-        Args:
-            device_name (str): Device name. Should follow the pattern
-                '/{device type}:{device number}' where device type is one of
-                CPU or GPU.
-        """
-        if not ALLOW_SWITCHERS and device_name != DEVICES['DEFAULT']:  # pragma: no cover
-            # no testing is implemented for warnings
-            warnings.warn("Device should not be changed after allocating gates.",
-                          category=RuntimeWarning)
-        parts = device_name[1:].split(":")
-        if device_name[0] != "/" or len(parts) < 2 or len(parts) > 3:
-            raise_error(ValueError, "Device name should follow the pattern: "
-                             "/{device type}:{device number}.")
-        device_type, device_number = parts[-2], int(parts[-1])
-        if device_type not in {"CPU", "GPU"}:
-            raise_error(ValueError, f"Unknown device type {device_type}.")
-        if device_number >= len(DEVICES[device_type]):
-            raise_error(ValueError, f"Device {device_name} does not exist.")
-
-        DEVICES['DEFAULT'] = device_name
-        with tf.device(device_name):
-            tfmatrices.allocate_matrices()
-
-    def get_device():
-        """Get execution device.
-
-        Returns:
-            A string with the device name.
-        """
-        return DEVICES['DEFAULT']
-
 
 else: # pragma: no cover
     # case not tested because the backend is preset to TensorFlow
